+++ readable/src/commonMain/kotlin/io/github/aughtone/readable/ordinality/Int.functions.kt
@@ -1,7 +1,6 @@
 package io.github.aughtone.readable.ordinality
 
 import io.github.aughtone.types.locale.Locale
-import io.github.aughtone.types.locale.currentNativeLocale
 
 /**
  * Formats a number into its ordinal string representation (e.g., 1 to "1st").
@@ -10,6 +9,6 @@ import io.github.aughtone.types.locale.currentNativeLocale
  * @param locale The locale to use for formatting rules.
  * @return The ordinal string representation of the number.
  */
-fun formatOrdinal(number: Int, locale: Locale = currentNativeLocale()): String {
+fun formatOrdinal(number: Int, locale: Locale = Locale.current): String {
     return ordinalityFor(locale)(number.toLong())
 }