+++ readable/src/commonMain/kotlin/io/github/aughtone/readable/ordinality/Long.extensions.kt
@@ -1,7 +1,6 @@
 package io.github.aughtone.readable.ordinality
 
 import io.github.aughtone.types.locale.Locale
-import io.github.aughtone.types.locale.currentNativeLocale
 
 /**
  * Extension function to format a [Long] into its ordinal string representation.
@@ -9,6 +8,6 @@ import io.github.aughtone.types.locale.currentNativeLocale
  * @param locale The locale to use for formatting rules.
  * @return The ordinal string representation of the number.
  */
-fun Long.toReadableOrdinal(locale: Locale = currentNativeLocale()): String {
+fun Long.toReadableOrdinal(locale: Locale = Locale.current): String {
     return formatOrdinal(this, locale)
 }